refactor: 취향항목 전체조회 API 레이어 분리 ( router / service / repository )
