+++ app/repositories/preference_repo.py
@@ -1,14 +1,14 @@
 from collections import defaultdict
 from typing import List
 
-from sqlalchemy.orm import Session
+from sqlalchemy.orm.session import Session
 
 from app.core.exception import UnknownExceptionError
 from app.model.users import Preferences
 from app.schema.preferences import PreferenceResponseModel, PreferenceType
 
 
-class PreferenceService:
+class PreferenceRepository:
     def __init__(self, db: Session) -> None:
         self.db = db
 
@@ -32,19 +32,5 @@ class PreferenceService:
         try:
             res = self.db.query(Preferences).all()
             return self.convert_list_to_model(res)
-
-        except Exception as e:
-            raise UnknownExceptionError(str(e))
-
-    async def get_preference_option(self, option_type: str):
-        """취향항목 조회하는 메소드."""
-        try:
-            res = (
-                self.db.query(Preferences.id, Preferences.name)
-                .filter(Preferences.type == option_type)
-                .all()
-            )
-            return [PreferenceType(id=r.id, name=r.name) for r in res]
-
         except Exception as e:
             raise UnknownExceptionError(str(e))
